Changed get_field_label so if verbose_name is set it displays as set
Previously if verbose_name for a field was  set to something like:
'Verbose Name' it would be changed to 'Verbose name' from .capitalize(),
this update changes the behavior so it will remain as is. Now it only
capitalizes first letter of verbose_name without capitalizing or making
other letters lowercase.