Field template fix for 2+ views that use the same mapping
This fixes an issue for SeekerViews that use the same document mapping
but different SeekerViews and have custom field template overrides on
one or more of the fields. This fix adds the optional variable
view_name, which is to be used as a unique name in a project that has
multiple SeekerViews that use the same document (mapping). If this is
not set the method get_view_name() will return the subclass name and
document name. In most instances the SeekerViews with the same mapping
will be subclasses with different subclass names.