Fix for issues with result sets that are too large for elasticsearch
Previously seeker would show last page even if clicking on that page
would result in a TransportError, now the behavior has changed that if
the result set is larger than the max results you can browse in a sorted
elasticsearch search query, (the default value for this is 10000) The
last page is no longer shown. Arrows are also added in to page by 1. If
the user views the last available page they are notified that there are
more results they can view from the exported results.