Add model name class in addition to doc class for seeker views
Add warnings if model is not defined on mapping class.
Adding back the model name class will provide backwards compatibility
with old templates since that is how they were set before the change to
doc_type in seeker and elasticsearch 6. As doc_type is set to eventually
be removed in a future version of elasticsearch setting up this
functionality now should help ease that transition. Note that if you
have set up custom column templates that used something similar to
result.meta.doc_type only you will need to add in the model_lower as
well.