Add is_dynamic option to seeker view
This allows for easier ajax and js interaction with seeker template.
Add option for hidden columns list allowing column data to go to seeker
but not show in the template under optional columns (thus it can be used
in other column templates)
Change form_template to refer to the form instead of the save_form and
add save_form_template for save_form.
Add Changelog to track changes to seeker